% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{bt}
\alias{bt}
\title{Validate whether values fall in the ranges}
\usage{
bt(x, range)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{range}{A vector of ranges.}
}
\description{
Validate whether values fall in the ranges
}
\details{
"b/t" is short for "between". To achieve faster performance, one
could do \code{dplyr::between(x, lower(rng), upper(rng))}.
}
\examples{
rng <- tie(lower = rnorm(10), upper = rnorm(10) + 5)
bt(0.2017, rng)

}
