% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{lower}
\alias{lower}
\alias{upper}
\alias{level}
\alias{is_range}
\title{Helpers for "range"}
\usage{
lower(x)

upper(x)

level(x)

is_range(x)
}
\arguments{
\item{x}{A "range" object.}
}
\description{
Helpers for "range"
}
