% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{tie}
\alias{tie}
\title{Construct ranges}
\usage{
tie(lower, upper, level = NA_integer_)
}
\arguments{
\item{lower, upper}{lower and upper limits}

\item{level}{between 0 and 100}
}
\description{
Construct ranges
}
