% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tbl_sum}
\alias{reexports}
\alias{type_sum}
\alias{reexports}
\alias{obj_sum}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{pillar}{\code{\link[pillar]{type_sum}}, \code{\link[pillar]{obj_sum}}}

  \item{tibble}{\code{\link[tibble]{tbl_sum}}}
}}

