% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.R
\name{tie}
\alias{tie}
\title{Construct ranges}
\usage{
tie(lower, upper, level = NULL)
}
\arguments{
\item{lower, upper}{Numerics for lower and upper limits.}

\item{level}{Default \code{NULL} or numbers between 0 and 100.}
}
\value{
A "range" object
}
\description{
Construct ranges
}
\examples{
tie(lower = rnorm(10), upper = rnorm(10) + 5, level = 95L)

}
